Prepare Responder & Request for new termination logic
